split string tests into separate source files for each type
