+++ src/libraries/System.Security.Cryptography.Encoding/tests/System.Security.Cryptography.Encoding.Tests.csproj
@@ -18,7 +18,8 @@
   <Import Project="Cbor\System.Formats.Cbor.Shared.projitems" />
   <ItemGroup>
     <Compile Include="Cbor.Tests\CborReaderTests.cs" />
-    <Compile Include="Cbor.Tests\CborReaderTests.String.cs" />
+    <Compile Include="Cbor.Tests\CborReaderTests.ByteString.cs" />
+    <Compile Include="Cbor.Tests\CborReaderTests.TextString.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.Integer.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.Helpers.cs" />
     <Compile Include="Cbor.Tests\CborReaderTests.Map.cs" />
@@ -33,7 +34,8 @@
     <Compile Include="Cbor.Tests\CborWriterTests.Simple.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Tag.cs" />
     <Compile Include="Cbor.Tests\CborWriterTests.Integer.cs" />
-    <Compile Include="Cbor.Tests\CborWriterTests.String.cs" />
+    <Compile Include="Cbor.Tests\CborWriterTests.ByteString.cs" />
+    <Compile Include="Cbor.Tests\CborWriterTests.TextString.cs" />
     <Compile Include="Cbor.Tests\CborRoundtripTests.cs" />
     <Compile Include="Cbor.Tests\CoseKeyHelpers.cs" />
   </ItemGroup>
